// webapp/ui/src/app/api-client/models/lottery-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface LotteryResponse {
  close_reason?: (string | null);
  created_at: string;
  created_by_user_id: number;
  end_date: (string | null);
  id: number;
  initialize_abandoned_at?: (string | null);
  initialize_abandoned_error?: (string | null);
  is_offchain_vrf?: boolean;
  lottery_type?: string;
  max_total?: (number | null);
  name: string;
  proceeding_purchases_started_at?: (string | null);
  randomness_account?: (string | null);
  second_phase_started_at?: (string | null);
  status: string;
  vrf_seed?: (string | null);
}